+++ src/index.ts
@@ -2,10 +2,12 @@ export { default as EventSubscriber } from './EventSubscriber';
 export { default as JwtCredentialsManager } from './JwtCredentialsManager';
 export { default as RedisSubscriber } from './RedisSubscriber';
 export { default as SubscriptionServer } from './SubscriptionServer';
-
 export { AsyncQueue } from './AsyncUtils';
-export { CreateValidationRules } from './AuthorizedSocketConnection';
-export { CredentialsManager } from './CredentialsManager';
-export { Logger, CreateLogger, LogLevels } from './Logger';
-export { Subscriber } from './Subscriber';
-export { SubscriptionServerConfig } from './SubscriptionServer';
+
+export type { CreateValidationRules } from './AuthorizedSocketConnection';
+export type { CredentialsManager } from './CredentialsManager';
+export type { Logger, CreateLogger, LogLevels } from './Logger';
+export type { Subscriber } from './Subscriber';
+export type { SubscriptionServerConfig } from './SubscriptionServer';
+
+export type { JwtCredentials } from './JwtCredentialsManager';
